Increase resource requirements
I had to increase resource requirements substantially to finish a
recent MSSPE run. Ideally we might have different requirements for
MSSPE and ARTIC runs. I don't have time to merge this to master at the
moment, but I am commiting to a separate branch for reference when I
circle back to this